rename: 컴포넌트의 파일명을 컴포넌트명/index.tsx 에서 컴포넌트명.tsx 파일로 수정 (#104)
